Migrate API to utilize new SQL backend
This is a big change; for REST requests, there are several endpoint names to correct which are no longer pluralized. Additionally, we no longer need to wrap the data in our emissions within an object because the back end handles associations and is only looking for a single identifier or field.