Change username in QuestionForm to fetch user's existing display name
This changes the authority of the username state from Lexsur to QuestionForm. This is required -as far as I can tell- because a change in state of the child component does not trigger a re-render as the child is a class and must therefore manage its own state. Additionally, Lexsur is no longer rendered if an auth token is not present.