Add option for users to join Lexsurs as guests
This is sort of tricky. Like before, we generate a temporary user in order to minimize necessary changes to pre-existing code. Anon users work just like regular users except that they cannot access the home page or Dashboard as they are logged out. It is extremely important that server-side validation ensures anonymous users do not interface with REST services!!!!!!!!!