Add administrative view and mobx store
Sorry for this commit message this was a huge endeavor over several sleepless days and the changes have become too intermingled to conveniently separate

In a nutshell, this enables Lexroom which was previously disabled by a flag. Lexroom allows us to have an underlying Lexsur instance (eventually) but toggle which views exist. Our first view distinction is an admin view(incomplete) and a client view